Buffer file writer
In HolyC, to implement a buffered file writer based on an interface and a file class, I had to define both the interface and the class, including methods for opening, writing to, and closing the file, as well as buffering the write operations to improve performance. 

However, HolyC, being unique to TempleOS, doesn't follow the same object-oriented programming (OOP) paradigms found in languages like C++ or Java, particularly in the use of interfaces in the way you might expect from those languages. This is my conceptual approach to creating a buffered file writer for HolyC.